requirements.txt
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
flake8==7.1.1
flake8-docstrings==1.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
httplib2==0.31.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
pyparsing==3.3.1
pyTelegramBotAPI==4.22.1
pytest==8.3.3
pytest-timeout==2.3.1
python-dotenv==1.0.1
redis==7.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
snowballstemmer==3.0.1
urllib3==2.5.0
